requirements.txt
asgiref==3.3.1
backports.zoneinfo==0.2.1
boto3==1.17.17
botocore==1.20.17
certifi==2020.12.5
dj-database-url==0.5.0
Django==3.1.6
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-nested-admin==3.3.3
django-storages==1.11.1
docxtpl==0.11.3
et-xmlfile==1.0.1
fontawesome-free==5.15.2
gunicorn==20.0.4
jdcal==1.4.1
Jinja2==2.11.3
jmespath==0.10.0
jplephem==2.15
lunardate==0.2.0
lxml==4.6.2
MarkupSafe==1.1.1
numpy==1.20.1
openpyxl==3.0.6
psycopg2==2.8.6
pyCalverter==1.6.1
pyluach==1.2.1
python-dateutil==2.8.1
python-docx==0.8.10
python-monkey-business==1.0.0
pytz==2021.1
s3transfer==0.3.4
sgp4==2.17
six==1.15.0
skyfield==1.37
skyfield-data==3.0.0
sqlparse==0.4.1
tzdata==2021.1
urllib3==1.26.3
whitenoise==5.2.0
workalendar==15.0.0
